requirements.txt
asgiref==3.4.1
astroid==2.8.0
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.9.0
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.20.2
drf-spectacular-sidecar==2021.10.18
filelock==3.3.1
future==0.18.2
gunicorn==20.1.0
identify==2.3.1
idna==2.7
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
jsonschema==4.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
paypalrestsdk==1.13.1
platformdirs==2.4.0
pre-commit==2.15.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.11.0
PyJWT==2.3.0
pylint==2.11.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0
regex==2021.9.24
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
stripe==2.60.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.3.0
wrapt==1.12.1
Pillow
